chore: move loggerinterface to types package
