+++ pkg/util/log/logger.go
@@ -7,70 +7,11 @@ package log
 
 import (
 	"github.com/cihub/seelog"
+
+	"github.com/DataDog/datadog-agent/pkg/util/log/types"
 )
 
-// LoggerInterface provides basic logging methods.
-type LoggerInterface interface {
-	// Tracef formats message according to format specifier
-	// and writes to log with level = Trace.
-	Tracef(format string, params ...interface{})
-
-	// Debugf formats message according to format specifier
-	// and writes to log with level = Debug.
-	Debugf(format string, params ...interface{})
-
-	// Infof formats message according to format specifier
-	// and writes to log with level = Info.
-	Infof(format string, params ...interface{})
-
-	// Warnf formats message according to format specifier
-	// and writes to log with level = Warn.
-	Warnf(format string, params ...interface{}) error
-
-	// Errorf formats message according to format specifier
-	// and writes to log with level = Error.
-	Errorf(format string, params ...interface{}) error
-
-	// Criticalf formats message according to format specifier
-	// and writes to log with level = Critical.
-	Criticalf(format string, params ...interface{}) error
-
-	// Trace formats message using the default formats for its operands
-	// and writes to log with level = Trace
-	Trace(v ...interface{})
-
-	// Debug formats message using the default formats for its operands
-	// and writes to log with level = Debug
-	Debug(v ...interface{})
-
-	// Info formats message using the default formats for its operands
-	// and writes to log with level = Info
-	Info(v ...interface{})
-
-	// Warn formats message using the default formats for its operands
-	// and writes to log with level = Warn
-	Warn(v ...interface{}) error
-
-	// Error formats message using the default formats for its operands
-	// and writes to log with level = Error
-	Error(v ...interface{}) error
-
-	// Critical formats message using the default formats for its operands
-	// and writes to log with level = Critical
-	Critical(v ...interface{}) error
-
-	// Close flushes all the messages in the logger and closes it. It cannot be used after this operation.
-	Close()
-
-	// Flush flushes all the messages in the logger.
-	Flush()
-
-	// SetAdditionalStackDepth sets the additional number of frames to skip by runtime.Caller
-	SetAdditionalStackDepth(depth int) error
-
-	// Sets logger context that can be used in formatter funcs and custom receivers
-	SetContext(context interface{})
-}
+type LoggerInterface = types.LoggerInterface
 
 // Default returns a default logger
 func Default() LoggerInterface {
